ARM: dts: imx23-sansa: Remove 'enable-active-low' property
Property 'enable-active-low' does not exist. Only 'enable-active-high' is
valid, and when this property is absent the gpio regulator will act as
active low by default.

So remove the unexisting 'enable-active-low' property.

Currently the GPIO flag is 0, which means active-high. In order to make
the dts description accurate, pass the GPIO_ACTIVE_LOW flag instead.

This change is safe because the gpio regulator driver does not take the
GPIO flag polarity into account.

Signed-off-by: Fabio Estevam <fabio.estevam@nxp.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 71ba49902d35d025b3810d6f19028eff624c67a7 ]