ARM: dts: sun6i: a31: Use new sun6i-a31-r-intc compatible for NMI/R_INTC
We introduced a new compatible for the NMI or R_INTC interrupt
controller. This new compatible has the register region aligned
to the boundary listed in the SoC's memory map.

This patch converts the NMI/R_INTC node to using the new compatible,
and fixes up the register region and device node name.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Acked-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: 626c0a0ee921a3ea001e4b4f35db89186649f822 ]