Merge branch 'net-Introduction-of-the-tc-tests'
Lucas Bates says:

====================
net: Introduction of the tc tests

Apologies for sending this as one big patch. I've been sitting on this a little
too long, but it's ready and I wanted to get it out.

There are a limited number of tests to start - I plan to add more on a regular
basis.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 708d32e4e5ef025ff3d6d5610876f4d39f62865d ]